pyproject.toml
[tool.black]
line-length = 120
target-version = ['py37']

[tool.mypy]
ignore_missing_imports = true
namespace_packages = true
explicit_package_bases = true

[tool.pylint.'MESSAGES CONTROL']
disable = "C0103, E0401, W0212, R0903, C0116"
max-line-length = 120

[tool.poetry]
name = "metrics"
version = "0.1.0"
description = "Verifa metrics dashboard"
authors = ["Jacob Larfors <jarfors@verifa.io>"]
license = "Apache v2.0"

[tool.poetry.extras]
pandas = ["dash"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.4.0"
dash = "^2.3.1"
gunicorn = "^20.1.0"
tempo-api-python-client = "^0.5.1"
setuptools = "^65.4.0"
types-requests = "^2.28.0"
google-api-python-client = "^2.68.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.7.1"

[tool.poetry.dev-dependencies]
pylint = "^2.13.4"
mypy = "^0.942"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
